refactor CatalogOperationDispatcher
