+++ core/src/main/java/com/datastrato/gravitino/catalog/FilesetOperationDispatcher.java
@@ -1,0 +1,130 @@
+/*
+ * Copyright 2024 Datastrato Pvt Ltd.
+ * This software is licensed under the Apache License version 2.
+ */
+package com.datastrato.gravitino.catalog;
+
+import static com.datastrato.gravitino.catalog.PropertiesMetadataHelpers.validatePropertyForCreate;
+
+import com.datastrato.gravitino.EntityStore;
+import com.datastrato.gravitino.NameIdentifier;
+import com.datastrato.gravitino.Namespace;
+import com.datastrato.gravitino.StringIdentifier;
+import com.datastrato.gravitino.connector.HasPropertyMetadata;
+import com.datastrato.gravitino.exceptions.FilesetAlreadyExistsException;
+import com.datastrato.gravitino.exceptions.NoSuchFilesetException;
+import com.datastrato.gravitino.exceptions.NoSuchSchemaException;
+import com.datastrato.gravitino.exceptions.NonEmptyEntityException;
+import com.datastrato.gravitino.file.Fileset;
+import com.datastrato.gravitino.file.FilesetCatalog;
+import com.datastrato.gravitino.file.FilesetChange;
+import com.datastrato.gravitino.storage.IdGenerator;
+import java.util.Map;
+
+public class FilesetOperationDispatcher extends CatalogOperationDispatcher
+    implements FilesetCatalog {
+  /**
+   * Creates a new CatalogOperationDispatcher instance.
+   *
+   * @param catalogManager The CatalogManager instance to be used for catalog operations.
+   * @param store The EntityStore instance to be used for catalog operations.
+   * @param idGenerator The IdGenerator instance to be used for catalog operations.
+   */
+  public FilesetOperationDispatcher(
+      CatalogManager catalogManager, EntityStore store, IdGenerator idGenerator) {
+    super(catalogManager, store, idGenerator);
+  }
+
+  @Override
+  public NameIdentifier[] listFilesets(Namespace namespace) throws NoSuchSchemaException {
+    return doWithCatalog(
+        getCatalogIdentifier(NameIdentifier.of(namespace.levels())),
+        c -> c.doWithFilesetOps(f -> f.listFilesets(namespace)),
+        NoSuchSchemaException.class);
+  }
+
+  @Override
+  public Fileset loadFileset(NameIdentifier ident) throws NoSuchFilesetException {
+    NameIdentifier catalogIdent = getCatalogIdentifier(ident);
+    Fileset fileset =
+        doWithCatalog(
+            catalogIdent,
+            c -> c.doWithFilesetOps(f -> f.loadFileset(ident)),
+            NoSuchFilesetException.class);
+
+    // Currently we only support maintaining the Fileset in the Gravitino's store.
+    return EntityCombinedFileset.of(fileset)
+        .withHiddenPropertiesSet(
+            getHiddenPropertyNames(
+                catalogIdent,
+                HasPropertyMetadata::filesetPropertiesMetadata,
+                fileset.properties()));
+  }
+
+  @Override
+  public Fileset createFileset(
+      NameIdentifier ident,
+      String comment,
+      Fileset.Type type,
+      String storageLocation,
+      Map<String, String> properties)
+      throws NoSuchSchemaException, FilesetAlreadyExistsException {
+    NameIdentifier catalogIdent = getCatalogIdentifier(ident);
+    doWithCatalog(
+        catalogIdent,
+        c ->
+            c.doWithPropertiesMeta(
+                p -> {
+                  validatePropertyForCreate(p.filesetPropertiesMetadata(), properties);
+                  return null;
+                }),
+        IllegalArgumentException.class);
+    long uid = idGenerator.nextId();
+    StringIdentifier stringId = StringIdentifier.fromId(uid);
+    Map<String, String> updatedProperties =
+        StringIdentifier.newPropertiesWithId(stringId, properties);
+
+    Fileset createdFileset =
+        doWithCatalog(
+            catalogIdent,
+            c ->
+                c.doWithFilesetOps(
+                    f -> f.createFileset(ident, comment, type, storageLocation, updatedProperties)),
+            NoSuchSchemaException.class,
+            FilesetAlreadyExistsException.class);
+    return EntityCombinedFileset.of(createdFileset)
+        .withHiddenPropertiesSet(
+            getHiddenPropertyNames(
+                catalogIdent,
+                HasPropertyMetadata::filesetPropertiesMetadata,
+                createdFileset.properties()));
+  }
+
+  @Override
+  public Fileset alterFileset(NameIdentifier ident, FilesetChange... changes)
+      throws NoSuchFilesetException, IllegalArgumentException {
+    validateAlterProperties(ident, HasPropertyMetadata::filesetPropertiesMetadata, changes);
+
+    NameIdentifier catalogIdent = getCatalogIdentifier(ident);
+    Fileset alteredFileset =
+        doWithCatalog(
+            catalogIdent,
+            c -> c.doWithFilesetOps(f -> f.alterFileset(ident, changes)),
+            NoSuchFilesetException.class,
+            IllegalArgumentException.class);
+    return EntityCombinedFileset.of(alteredFileset)
+        .withHiddenPropertiesSet(
+            getHiddenPropertyNames(
+                catalogIdent,
+                HasPropertyMetadata::filesetPropertiesMetadata,
+                alteredFileset.properties()));
+  }
+
+  @Override
+  public boolean dropFileset(NameIdentifier ident) {
+    return doWithCatalog(
+        getCatalogIdentifier(ident),
+        c -> c.doWithFilesetOps(f -> f.dropFileset(ident)),
+        NonEmptyEntityException.class);
+  }
+}
